+++ CHANGELOG.rst
@@ -34,6 +34,8 @@ Fixed
 - Fixed ``'Namespace' object has no attribute 'persist_nlu_data'`` error during
   interactive learning
 - Pinned `networkx~=2.3.0` to fix visualization in `rasa interactive` and Rasa X
+- Fixed ``No model found`` error when using ``rasa run actions`` with "actions"
+  as a directory.
 
 [1.4.1] - 2019-10-22
 ^^^^^^^^^^^^^^^^^^^^