+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Rasa Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning`_ starting with version 1.0.
 
-[Unreleased 1.4.3]
+[Unreleased 1.4.4]
 ^^^^^^^^^^^^^^^^^^
 
 Added
@@ -18,6 +18,12 @@ Changed
 
 Removed
 -------
+
+Fixed
+-----
+
+[1.4.3] - 2019-10-29
+^^^^^^^^^^^^^^^^^^^^
 
 Fixed
 -----